Add docker image
Add Dockerfile to run `kg-statistics`. Right now there is an
error when running the image since the nginx configuration is
off. Moreover I had to install the node package `natives` for
gulp to run. This seems ok since the node version is pretty old
and fixed for this project. Also I moved the build result into a `build`
folder instead of `../ui`.